Improve reference to `ManagedBass.Midi`
Fixes #4

Improves the reference to `ManagedBass.Midi` to make it easier to run the project from source.

The issue was that `ppy.osu.Framework` uses a custom fork of `ManagedBass` (`ppy.ManagedBass`) which does have a corresponding `ManagedBass.Midi` package. This causes a conflict when trying to use `ManagedBass.Midi` in the same project as `ppy.osu.Framework` because `ManagedBass.Midi` requires types from the original `ManagedBass` package, not the `ppy.ManagedBass` package.

The original workaround was to clone [the repository](https://github.com/ppy/osu-framework) of the `ppy.osu.Framework` package, swap the reference to `ppy.ManagedBass` to `ManagedBass` (located [here](https://github.com/ppy/osu-framework/blob/451f8b2d614c906733dd74a94c04e5e044cbf3c2/osu.Framework/osu.Framework.csproj#L27C5-L27C73)), and use this local version of the framework when building the game. This is not ideal because it requires cloning the entire osu framework repository, modifying the framework code, and adding the osu framework project to the game's solution.

Due to the way that transitive dependencies are handled, there are no easy ways to resolve this issue. On top of that, the names of the types or namespaces were not changed during the forking process so there are conflicts everywhere (a type with the same name will be present in the exact same namespace in two assemblies). Even the concept of assembly binding redirects does not work in this case because the new version (`ppy.ManagedBass`) has a different public key token and name than the original version (`ManagedBass`).

To resolve this, an explicit reference to the `ManagedBass` package was added and it was aliased to `OriginalManagedBass`. By using an alias, `ManagedBass` in the code with `ManagedBass.Midi`, and the osu framework can continue to use the `ppy.ManagedBass` package.

This is still not ideal because the code that uses `ManagedBass` will need to use the alias `OriginalManagedBass` to access the types from the original `ManagedBass` package. However, this is a much simpler solution than the previous workaround.